+++ crates/chia-sdk-driver/src/clear_signing.rs
@@ -2,6 +2,7 @@ mod facts;
 mod inner_spend;
 mod linked_spend;
 mod requested_payments;
+mod signing;
 mod vault;
 mod vault_message;
 mod vault_transaction;
@@ -10,6 +11,7 @@ pub use facts::*;
 pub use inner_spend::*;
 pub use linked_spend::*;
 pub use requested_payments::*;
+pub use signing::*;
 pub use vault::*;
 pub use vault_message::*;
 pub use vault_transaction::*;
